Release 9.0

See merge request octopus-code/octopus!506

+++ liboct_parser/parser_f.c
@@ -29,6 +29,7 @@
 #include <string.h>
 #include <math.h>
 #include <assert.h>
+#include <stdint.h>
 
 #include "liboct_parser.h"
 #include "symbols.h"
@@ -154,7 +155,7 @@ int FC_FUNC_(oct_parse_isdef, OCT_PARSE_ISDEF)
 
 /* --------------------------------------------------------- */
 void FC_FUNC_(oct_parse_int, OCT_PARSE_INT)
-	(STR_F_TYPE name, int *def, int *res STR_ARG1)
+	(STR_F_TYPE name, int64_t *def, int64_t *res STR_ARG1)
 { 
   char *name_c;
